requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
SQLAlchemy==2.0.30
python-dotenv==1.0.1
flask-jwt-extended==4.6.0
Flask-SQLAlchemy==3.1.1
python-dateutil==2.9.0